# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hubuum REST API",
    "description": "OpenAPI documentation for the Hubuum REST service.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment base URL"
    }
  ],
  "paths": {
...
    "/api/v1/iam/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Api V1 Iam Users",
        "description": "Auto-generated documentation for GET /api/v1/iam/users. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1IamUsers",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users matching optional query filters The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Post Api V1 Iam Users",
        "description": "Auto-generated documentation for POST /api/v1/iam/users.",
        "operationId": "postApiV1IamUsers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/iam/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Api V1 Iam Users By User Id",
        "description": "Auto-generated documentation for GET /api/v1/iam/users/{user_id}.",
        "operationId": "getApiV1IamUsersByUserId",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete Api V1 Iam Users By User Id",
        "description": "Auto-generated documentation for DELETE /api/v1/iam/users/{user_id}.",
        "operationId": "deleteApiV1IamUsersByUserId",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Patch Api V1 Iam Users By User Id",
        "description": "Auto-generated documentation for PATCH /api/v1/iam/users/{user_id}.",
        "operationId": "patchApiV1IamUsersByUserId",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/iam/users/{user_id}/groups": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Api V1 Iam Users By User Id Groups",
        "description": "Auto-generated documentation for GET /api/v1/iam/users/{user_id}/groups. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1IamUsersByUserIdGroups",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groups of user The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/iam/users/{user_id}/tokens": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Api V1 Iam Users By User Id Tokens",
        "description": "Auto-generated documentation for GET /api/v1/iam/users/{user_id}/tokens. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1IamUsersByUserIdTokens",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active tokens for user The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserToken"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace_id}/permissions/user/{user_id}": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "List all permissions for a user on a namespace",
        "description": "Auto-generated documentation for GET /api/v1/namespaces/{namespace_id}/permissions/user/{user_id}. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1NamespacesByNamespaceIdPermissionsUserByUserId",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User permissions via group memberships The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupPermission"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No permissions found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "error": "Unauthorized",
          "message": "Authentication failure"
        }
      },
...
      "Group": {
        "type": "object",
        "required": [
          "id",
          "groupname",
          "description",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "groupname": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GroupPermission": {
        "type": "object",
        "required": [
          "group",
          "permission"
        ],
        "properties": {
          "group": {
            "$ref": "#/components/schemas/Group"
          },
          "permission": {
            "$ref": "#/components/schemas/Permission"
          }
        }
      },
...
      "NewUser": {
        "type": "object",
        "description": "Struct to create a new user.\n\nThe password is expected to be hashed\nbefore being passed to the database.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "example": {
          "email": "alice@example.com",
          "password": "correct-horse-battery-staple",
          "username": "alice"
        }
      },
...
      "Permission": {
        "type": "object",
        "required": [
          "id",
          "namespace_id",
          "group_id",
          "has_read_namespace",
          "has_update_namespace",
          "has_delete_namespace",
          "has_delegate_namespace",
          "has_create_class",
          "has_read_class",
          "has_update_class",
          "has_delete_class",
          "has_create_object",
          "has_read_object",
          "has_update_object",
          "has_delete_object",
          "has_create_class_relation",
          "has_read_class_relation",
          "has_update_class_relation",
          "has_delete_class_relation",
          "has_create_object_relation",
          "has_read_object_relation",
          "has_update_object_relation",
          "has_delete_object_relation",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "group_id": {
            "type": "integer",
            "format": "int32"
          },
          "has_create_class": {
            "type": "boolean"
          },
          "has_create_class_relation": {
            "type": "boolean"
          },
          "has_create_object": {
            "type": "boolean"
          },
          "has_create_object_relation": {
            "type": "boolean"
          },
          "has_delegate_namespace": {
            "type": "boolean"
          },
          "has_delete_class": {
            "type": "boolean"
          },
          "has_delete_class_relation": {
            "type": "boolean"
          },
          "has_delete_namespace": {
            "type": "boolean"
          },
          "has_delete_object": {
            "type": "boolean"
          },
          "has_delete_object_relation": {
            "type": "boolean"
          },
          "has_read_class": {
            "type": "boolean"
          },
          "has_read_class_relation": {
            "type": "boolean"
          },
          "has_read_namespace": {
            "type": "boolean"
          },
          "has_read_object": {
            "type": "boolean"
          },
          "has_read_object_relation": {
            "type": "boolean"
          },
          "has_update_class": {
            "type": "boolean"
          },
          "has_update_class_relation": {
            "type": "boolean"
          },
          "has_update_namespace": {
            "type": "boolean"
          },
          "has_update_object": {
            "type": "boolean"
          },
          "has_update_object_relation": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "namespace_id": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateUser": {
        "type": "object",
        "description": "Struct to update a user.\n\nThe password, if present, is expected to be hashed\nbefore being passed to the database.",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "example": {
          "email": "alice@example.com",
          "password": "new-password",
          "username": "alice"
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "password",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "password": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserToken": {
        "type": "object",
        "required": [
          "token",
          "user_id",
          "issued"
        ],
        "properties": {
          "issued": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}